some fix about contiguous

+++ methods/anyflow/far/utils/dist_util.py
@@ -78,7 +78,7 @@ def _fsdp1_wrap_npu(
 ):
     if mixed_precision:
         mixed_precision_policy = MixedPrecision(
-            param_dtype=torch.float16,
+            param_dtype=torch.bfloat16,
             reduce_dtype=torch.float32,
             buffer_dtype=torch.float32,
             cast_forward_inputs=False,
